backend/requirements.txt
fastapi
uvicorn
pydantic
requests
jwt
itsdangerous
google-cloud-storage
google-generativeai
Jinja2
pyyaml
psycopg2
google-api-python-client
